requirements.txt
streamlit==1.29.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
sentence-transformers==2.2.2
chromadb==0.4.18
opencv-python-headless==4.8.1